Internals

Internals

Lattice

HartreeFockBogoliubov.Lattice.CarteCoord — Type.

CarteCoord

Cartesian coordinates. Vector{Float64}.

source

HartreeFockBogoliubov.Lattice.FractCoord — Type.

FractCoord

Fractional coordinates.

Members

whole ::Vector{Int}: Integer part of fractional coordinates

fraction ::Vector{Float64}: [0,1) part of fractional coordinates

source

HartreeFockBogoliubov.Lattice.UnitCell — Type.

UnitCell{O}

Parameters

O: type of "orbital". Any type can be used, but we recommend using String or tuple of String and Int for compatibility with JSON.

Members

latticevectors ::Array{Float64, 2}: Lattice vectors

reducedreciprocallatticevectors ::Array{Float64, 2}: Reduced reciprocal lattice vectors (transpose of inverse of latticevectors)

reciprocallatticevectors ::Array{Float64, 2}: Reciprocal lattice vectors. 2π * reducedreciprocallatticevectors

orbitals ::Vector{Tuple{T, FractCoord}}: List of orbitals within unit cell

orbitalindices ::Dict{T, Int}: Indices of orbitals

source

HartreeFockBogoliubov.Lattice.addorbital! — Method.

addorbital!

Add an orbital to the unit cell.

Arguments

uc ::UnitCell{T}

orbitalname ::{T}

orbitalcoord ::FractCoord

source

HartreeFockBogoliubov.Lattice.carte2fract — Method.

carte2fract

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

cc ::CarteCoord: cartesian coordinates

tol ::Real=sqrt(eps(Float64)): tolerance

source

HartreeFockBogoliubov.Lattice.carte2fract — Method.

carte2fract

Arguments

latticevectors ::Array{Float64, 2}

cc ::CarteCoord

source

HartreeFockBogoliubov.Lattice.dimension — Method.

dimension

Dimension of the fractional coordinates

Arguments

fc ::FractCoord: Fractional coordinates.

source

HartreeFockBogoliubov.Lattice.dimension — Method.

dimension

Spatial dimension of the unit cell.

source

HartreeFockBogoliubov.Lattice.fract2carte — Method.

fract2carte

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: square matrix whose columns are lattice vectors.

fc ::FractCoord: fractional coordinates

source

HartreeFockBogoliubov.Lattice.fract2carte — Method.

fract2carte

Arguments

latticevectors ::Array{Float64, 2}

fc ::FractCoord

source

HartreeFockBogoliubov.Lattice.getorbital — Method.

getorbital

Arguments

uc ::UnitCell{T}

index ::Integer

Return

(orbitalname, fractcoord)

source

HartreeFockBogoliubov.Lattice.getorbital — Method.

getorbital

Get the orbital (its orbital name and its fractional coordinates) with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

(orbitalname, fractcoord)

source

HartreeFockBogoliubov.Lattice.getorbitalcoord — Method.

getorbitalcoord

Arguments

uc ::UnitCell

idx ::Integer

Return

fractcoord

source

HartreeFockBogoliubov.Lattice.getorbitalcoord — Method.

getorbitalcoord

Get the fractional coordinates of the orbital with the given name.

Arguments

uc ::UnitCell{O}

name ::O

Return

fractcoord

source

HartreeFockBogoliubov.Lattice.getorbitalindex — Method.

getorbitalindex

Get index of the given orbital.

Arguments

uc ::UnitCell{O}

name ::O

source

HartreeFockBogoliubov.Lattice.getorbitalindexcoord — Method.

getorbitalindexcoord

Arguments

uc ::UnitCell{T}

name ::T

Return

(index, fractcoord)

source

HartreeFockBogoliubov.Lattice.getorbitalname — Method.

getorbitalname

Arguments

uc ::UnitCell

index ::Integer

Return

orbitalname

source

HartreeFockBogoliubov.Lattice.hasorbital — Method.

hasorbital{T}

Test whether the unit cell contains the orbital of given name.

Arguments

uc ::UnitCell{O}

name ::O

source

HartreeFockBogoliubov.Lattice.linpath — Method.

momentumpath

Generate a list of momenta

Arguments

anchorpoints

(Optional) nseg - number of points in each segment

source

HartreeFockBogoliubov.Lattice.make_unitcell — Method.

UnitCell

Construct an n-dimensional lattice.

Arguments

latticevectors ::AbstractArray{<:AbstractFloat, 2}: Lattice vectors

OrbitalType::DataType

Optional Arguments

tol=sqrt(eps(Float64)): Epsilon

source

HartreeFockBogoliubov.Lattice.make_unitcell — Method.

UnitCell

Construct a one-dimensional lattice.

Arguments

latticeconstant ::Float64: Lattice constant

OrbitalType: List of orbitals

Optional Arguments

tol=sqrt(eps(Float64)): Tolerance

source

HartreeFockBogoliubov.Lattice.momentumgrid — Method.

momentumgrid

Generate an n-dimensional grid of momenta of given shape

source

HartreeFockBogoliubov.Lattice.momentumpath — Method.

momentumpath

The anchorpoints are given in units of the reciprocal lattice vectors.

source

HartreeFockBogoliubov.Lattice.numorbital — Method.

numorbital

Number of orbitals of the unit cell.

Arguments

uc ::UnitCell

source

HartreeFockBogoliubov.Lattice.whichunitcell — Method.

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

source

HartreeFockBogoliubov.Lattice.whichunitcell — Method.

whichunitcell

Return

R ::Vector{Int}: which unit cell the specificied orbital/cartesian coordinates belongs to.

source

Base.isapprox — Method.

isapprox(x, y; rtol::Real=atol>0 ? 0 : √eps, atol::Real=0, nans::Bool=false, norm::Function)

source

Spec

HartreeFockBogoliubov.Spec — Module.

Submodule `Spec`

source

HartreeFockBogoliubov.Spec.Hopping — Constant.

Hopping

source

HartreeFockBogoliubov.Spec.Interaction — Constant.

Interaction

source

HartreeFockBogoliubov.Spec.FullHamiltonian — Type.

FullHamiltonian

Members

unitcell ::UnitCell

hoppings ::Vector{Hopping}

interactions ::Vector{Interaction}

source

HartreeFockBogoliubov.Spec.FullHamiltonian — Method.

Hamiltonian

Create an empty Hamiltonian

Arguments

unitcell ::UnitCell

source

HartreeFockBogoliubov.Spec.HoppingDiagonal — Type.

HoppingDiagonal{R<:Real}

Represents

\[ t c_{i}^{*} c_{i}\]

Members

amplitude ::R

i ::Int: index of orbital

Ri ::Vector{Int}: which unit cell? (indexed by a1, and a2)

source

HartreeFockBogoliubov.Spec.HoppingDiagonal — Method.

source

HartreeFockBogoliubov.Spec.HoppingOffdiagonal — Type.

HoppingOffdiagonal{C<:Number}

Represents

\[ t c_{i}^{*} c_{j} + t^* c_{j}^{*} c_{i}\]

t, i, j and the unitcell-coordinates Ri and Rj are stored. Require that (i, Ri) <= (j, Rj)

Members

amplitude ::C

i, j ::T

Ri, Rj ::Vector{Int}

source

HartreeFockBogoliubov.Spec.InteractionDiagonal — Type.

InteractionDiagonal{R<:Real}

Represents

\[ U c_{i}^{*} c_{j}^{*} c_{j} c_{i}\]

Members

amplitude ::R

i, j ::Int

Ri, Rj ::Vector{Int}

source

HartreeFockBogoliubov.Spec.InteractionOffdiagonal — Type.

InteractionOffdiagonal{C<:Number}

i < j, k < l, i < k or (i == k and j < l)

Represents

\[ U c_{i}^{*} c_{j}^{*} c_{l} c_{k} + U^{*} c_{k}^{*} c_{l}^{*} c_{j} c_{i}\]

Only keep the first term (and require i < j, k < l, i <= k)

Ordering of orbitals by (i, Ri)

Members

amplitude ::C

i, j, k, l ::Int

Ri, Rj, Rk, Rl ::Vector{Int}

source

HartreeFockBogoliubov.Spec.addhopping! — Method.

addhopping!

Arguments

hamiltonian ::Hamiltonian

hopping ::HoppingOffdiagonal

source

HartreeFockBogoliubov.Spec.addhopping! — Method.

addhopping!

Arguments

hamiltonian ::Hamiltonian

hopping ::HoppingDiagonal

source

HartreeFockBogoliubov.Spec.addinteraction! — Method.

addinteraction!

Arguments

hamiltonian ::Hamiltonian

interaction ::InteractionOffdiagonal

source

HartreeFockBogoliubov.Spec.addinteraction! — Method.

addinteraction!

Arguments

hamiltonian ::Hamiltonian

interaction ::InteractionDiagonal

source

HartreeFockBogoliubov.Spec.hoppingbycarte — Method.

hoppingbycarte{T}

Make a hopping element with cartesian coordinates.

Arguments

uc ::UnitCell{T}

amplitude ::Number

i ::T

j ::T

ri ::CarteCoord

rj ::CarteCoord

tol ::Real : Optional. Defaults to sqrt(eps(Float64))

source

HartreeFockBogoliubov.Spec.hoppingbycarte — Method.

hoppingbycarte{T}

Make a hopping element with cartesian coordinates.

Arguments

uc ::UnitCell{T}

amplitude ::Real

i ::T

ri ::CarteCoord

tol ::Real : Optional. Defaults to sqrt(eps(Float64))

source

HartreeFockBogoliubov.Spec.interactionbycarte — Method.

interactionbycarte{T}

Make an interaction element with cartesian coordinates.

Arguments

* `uc ::UnitCell{T}`
* `amplitude ::Number`
* `i ::T`
* `j ::T`
* `k ::T`
* `l ::T`
* `ri ::CarteCoord`
* `rj ::CarteCoord`
* `rk ::CarteCoord`
* `rl ::CarteCoord`
* `tol ::Real` : Optional. Defaults to `sqrt(eps(Float64))`

source

HartreeFockBogoliubov.Spec.interactionbycarte — Method.

interactionbycarte{T}

Make an interaction element with cartesian coordinates.

Arguments

* `uc ::UnitCell{T}`
* `amplitude ::Number`
* `i ::T`
* `j ::T`
* `ri ::CarteCoord`
* `rj ::CarteCoord`
* `tol ::Real` : Optional. Defaults to `sqrt(eps(Float64))`

source

HartreeFockBogoliubov.Spec.islocal — Method.

islocal

Check if the hopping element is local (i.e. Ri is zero)

source

HartreeFockBogoliubov.Spec.localized — Method.

localized

Return a hopping element that is local (i.e. Ri is zero)

source

Generator

HartreeFockBogoliubov.Generator — Module.

Generator submodule

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

HartreeFockBogoliubov.Generator.hopping_inplace — Method.

hopping_inplace

source

Topology

HartreeFockBogoliubov.Topology.squarify — Method.

squarify

Arguments

uc::Spec.FullHamiltonian{O}

source

HartreeFockBogoliubov.Topology.squarify — Method.

squarify

Arguments

uc::HFB.HFBHamiltonian{O}

source

HartreeFockBogoliubov.Topology.squarify — Method.

squarify

In order to make the Hamiltoinian a periodic function of momentum, bring all the sites to the origin. In addition, make the unitcell into a square.

Arguments

uc::Lattice.UnitCell{O}

source

HartreeFockBogoliubov.Topology.chernnumber — Method.

chernnumber

Compute chern number of the band structure defined by the hoppings and the selected bands.

source

HartreeFockBogoliubov.Topology.z2index — Method.

z2index

Compute Z₂ index of time-reversal-invariant Hamiltonian.

Arguments

uc::UnitCell{O}

hops::AbstractVector{Hopping}

timereversal::AbstractMatrix

n1 ::Integer

n2 ::Integer

selectpairs::AbstractVector{<:Integer}

Optional Arguments

tol ::Real = sqrt(eps(Float64))

Returns

(The Z2 index, max| Hₖ - T⁻¹HₖT | for k in TRIMs)

source

HFB

HartreeFockBogoliubov.HFB.HFBComputer — Type.

HFBConmputer is a type holding the ρ, t and Γ, Δ of a Hartree-Fock-Bogoliubov Hamiltonian.

source

HartreeFockBogoliubov.HFB.HFBComputer — Method.

source

HartreeFockBogoliubov.HFB.HFBHamiltonian — Type.

source

HartreeFockBogoliubov.HFB.HFBHamiltonian — Method.

source

HartreeFockBogoliubov.HFB.HFBHamiltonian — Method.

source

HartreeFockBogoliubov.HFB.HFBSolver — Type.

source

HartreeFockBogoliubov.HFB.HFBSolver — Method.

source

HartreeFockBogoliubov.HFB.addinteraction! — Method.

Add offdiagonal interaction

source

HartreeFockBogoliubov.HFB.addinteraction! — Method.

addinteraction!

Add diagonal interaction

source

HartreeFockBogoliubov.HFB.compute_hfbfield! — Method.

Compute Γ and Δ from ρ and t.

source

HartreeFockBogoliubov.HFB.isvalid — Method.

source

HartreeFockBogoliubov.HFB.isvalid — Method.

source

HartreeFockBogoliubov.HFB.loop — Method.

loop

Perform selfconsistency loop a number of times with the given precondition and given update functions.

Arguments

solver ::HFBSolver{T}

sol::HFBAmplitude

run::Integer

Optional Arguments

update::Function=simpleupdate

precondition::Function=identity: on target field

callback::Function=_noop: Function called after every update as

callback(i, run)

source

HartreeFockBogoliubov.HFB.loop_threaded — Method.

loop

Perform selfconsistency loop a number of times with the given precondition and given update functions.

Arguments

solver ::HFBSolver{T}

sf::HFBAmplitude

run::Integer

Optional Arguments

update::Function=simpleupdate

precondition::Function=identity

callback::Function=_noop: Function called after every update as

callback(i, run)

source

HartreeFockBogoliubov.HFB.make_Deltamatrix — Method.

Return a generator of Δ matrix (which is a function of momentum)

source

HartreeFockBogoliubov.HFB.make_Gammamatrix — Method.

Return a generator of Γ matrix (which is a function of momentum)

source

HartreeFockBogoliubov.HFB.make_greencollector — Method.

make_greencollector

Returns a function which has the following signature

collector(k, eigenvalues, eigenvectors, ρout, tout)

source

HartreeFockBogoliubov.HFB.make_hamiltonian — Method.

source

HartreeFockBogoliubov.HFB.make_hfbamplitude — Method.

func : (idx, i, j, r) -> val

source

HartreeFockBogoliubov.HFB.make_hfbamplitude — Method.

func : (idx, i, j, r) -> 0

source

HartreeFockBogoliubov.HFB.make_hfbamplitude — Method.

Check if hint contains ρ

source

HartreeFockBogoliubov.HFB.make_hfbfield — Method.

Compute Γ and Δ from ρ and t.

source

HartreeFockBogoliubov.HFB.make_hfbfield — Method.

Compute Γ and Δ from ρ and t.

source

HartreeFockBogoliubov.HFB.make_hint — Method.

source

HartreeFockBogoliubov.HFB.make_hoppingmatrix — Method.

Return a generator of hopping matrix (which is a function of momentum)

source

HartreeFockBogoliubov.HFB.next_hfbamplitude! — Method.

source

HartreeFockBogoliubov.HFB.next_hfbamplitude — Method.

source

HartreeFockBogoliubov.HFB.next_hfbamplitude — Method.

source

HartreeFockBogoliubov.HFB.next_hfbamplitude_threaded! — Method.

source

HartreeFockBogoliubov.HFB.next_hfbamplitude_threaded — Method.

source

HartreeFockBogoliubov.HFB.randomize! — Method.

Randomize a hfbamplitude

source

HartreeFockBogoliubov.HFB.CollectRow — Type.

CollectRow is holds info on how to compute ρ or t. Its elements are:

Is diagonal? (only for rho)

row orbital

col orbital

displacement r(col) - r(row)

source

HartreeFockBogoliubov.HFB.DeployRow — Type.

DeployRow is holds info on how to compute Γ or Δ. Its elements are:

Is diagonal

row orbital

col orbital

displacement r(col) - r(row)

list of sources, each of which is a tuple of

index of ρ or t from which to compute this Γ or Δ.

amplitude (coefficient to multiply to ρ or t)

boolean indicating whether (1) conjugation is needed (for ρ/Γ) or (2) minus sign is needed (for t/Δ).

source

HartreeFockBogoliubov.HFB.HoppingMeanField — Type.

source

HartreeFockBogoliubov.HFB.PairingMeanField — Type.

source

HartreeFockBogoliubov.HFB.make_particleholeregistry — Method.

source

HartreeFockBogoliubov.HFB.make_particleparticleregistry — Method.

source

HartreeFockBogoliubov.HFB.next_hfbfield — Method.

source

LinearizedGap

HartreeFockBogoliubov.LinearizedGap.linearizedpairingkernel — Method.

linearizedpairingkernel

Compute the kernel Γ of the linearized gap equation in the pairing channel which is written as Δ = Γ⋅Δ

source